% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basymmetry.R
\name{basymmetry}
\alias{basymmetry}
\title{Asymmetry indices}
\usage{
basymmetry(filename, labelnametibble, option = 1)
}
\arguments{
\item{filename}{One or more NIfTI-files}

\item{labelnametibble}{A label name tibble with correct format and unique
columns. First column of the label name tibble should contain an integer
label number and second column should contain a label name string. Label
numbers and label names must be unique. The trailing characters of label
names must be R or L.}

\item{option}{1: Returns asymmetry indices for whole brain. 2: Returns
asymmetry indices for the left and right regions.}
}
\value{
Asymmetry index/indices for whole brain or asymmetry indices for
the left and right regions for supplied NIfTI-file(s).
}
\description{
Calculates asymmetry indices for one or more supplied
NIfTI-files.
}
\details{
Input one or more NIfTI-files and returns asymmetry indices for
whole brain or returns asymmetry indices for the left and right regions.
}
