% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basymmetry2.R
\name{basymmetry2}
\alias{basymmetry2}
\title{Asymmetry indices analysis data}
\usage{
basymmetry2(filename, labelnametibble)
}
\arguments{
\item{filename}{One or more NIfTI-files}

\item{labelnametibble}{A label name tibble with correct format and unique
columns. First column of the label name tibble should contain an integer
label number and second column should contain a label name string. Label
numbers and label names must be unique. The trailing characters of label
names must be R or L.}
}
\value{
Asymmetry index/indices for the left and right regions for supplied
NIfTI-file(s).
}
\description{
Calculates asymmetry indices for one or more supplied
NIfTI-files.
}
\details{
Inputs one or more NIfTI-files to calculate asymmetry indices for
one or more NIfTI-files.
}
