% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsimilarity.R
\name{bsimilarity}
\alias{bsimilarity}
\title{Dice coefficients and Jaccard indices}
\usage{
bsimilarity(filename1, filename2, labelnametibble)
}
\arguments{
\item{filename1}{One NIfTI-file}

\item{filename2}{One NIfTI-file}

\item{labelnametibble}{A label name tibble with correct format and unique
columns. First column of the label name tibble should contain an integer
label number and second column should contain a label name string. Label
numbers and label names must be unique.}
}
\value{
Dice coefficients and Jaccard indices for a supplied file pair.
Only regions which are in the label name tibble will be included in the
result.
}
\description{
Calculates Dice coefficients and Jaccard indices for a file
pair.
}
\details{
Function is used to calculate Dice coefficients and Jaccard indices
for a file pair. A file pair contains two segmented NIfTI -files which are
generated from the same source.
}
