% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsimilarity2.R
\name{bsimilarity2}
\alias{bsimilarity2}
\title{Jaccard indices analysis data}
\usage{
bsimilarity2(filename1, filename2, labelnametibble)
}
\arguments{
\item{filename1}{One or more NIfTI-files}

\item{filename2}{One or more NIfTI-files}

\item{labelnametibble}{A label name tibble with correct format and unique
columns. First column of the label name tibble should contain an integer
label number and second column should contain a label name string. Label
numbers and label names must be unique.}
}
\value{
Jaccard indices for the supplied file pairs.
}
\description{
Calculates Jaccard indices for one or more file pairs.
}
\details{
Function is used to calculate Jaccard indices for all regions for
one or more file pairs. Each file pair contains two segmented NIfTI -files
which are generated from the same source.
directory.
}
