% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bzscore.R
\name{bzscore}
\alias{bzscore}
\title{Region z-scores}
\usage{
bzscore(filename, labelnametibble, refstats)
}
\arguments{
\item{filename}{One or more NIfTI-files for individuals from which z-scores
are to be calculated}

\item{labelnametibble}{A label name tibble}

\item{refstats}{A tibble which was previously calculated by refmeanstd
function}
}
\value{
z-scores for all regions with label name for one or more NIfTI-files
}
\description{
Calculates z-scores of all regions.
}
\details{
Inputs one or more NIfTI-files from individuals, for which z-scores
are to be calculated, a label name tibble with correct formats and unique
columns and a tibble and a tibble including mean values and standard
deviations from the reference group for one or more individuals. Returns
z-scores, with label names, for all regions.
}
