% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bzscore2.R
\name{bzscore2}
\alias{bzscore2}
\title{z-scores analysis data}
\usage{
bzscore2(targets, labelnametibble, refstats)
}
\arguments{
\item{targets}{One or more NIfTI-files for individuals from which z-scores
are to be calculated}

\item{labelnametibble}{A label name tibble}

\item{refstats}{a tibble which was previously calculated by refmeanstd
function}
}
\value{
z-scores for all regions with label name for one or more NIfTI-files
}
\description{
Calculates and saves z-scores of all regions.
}
\details{
Input one or many NIfTI-files from individuals, for which z-scores
are to be calculated, a label name tibble with correct formats and unique
columns and a tibble including mean values and standard deviations from the
reference group for one or more individuals.
}
