% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_label.R
\name{check_label}
\alias{check_label}
\title{Labelnametibble check}
\usage{
check_label(labelnametibble)
}
\arguments{
\item{labelnametibble}{A label name tibble}
}
\value{
An error message or a label name tibble.
}
\description{
Checks the format of the label name tibble.
}
\details{
Checks if label number is an integer and if there are any duplicate
label numbers or label names in the tibble.
}
