% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_label_RGB.R
\name{check_label_RGB}
\alias{check_label_RGB}
\title{Check labelnametibble with R, G, B columns}
\usage{
check_label_RGB(labelnametibble)
}
\arguments{
\item{labelnametibble}{A label name tibble}
}
\value{
An error message or a label name tibble.
}
\description{
Checks the format of the label name tibble.
}
\details{
Checks if label number is an integer and if there are any duplicate
label numbers or label names in the tibble. Checks if the tibble includes R,
G, B columns.
}
