% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element_count.R
\name{element_count}
\alias{element_count}
\title{Counts for elements}
\usage{
element_count(slice_no, targetlabel, slice_img1, slice_img2)
}
\arguments{
\item{slice_no}{slice_no with integer format}

\item{targetlabel}{Label_No with integer format}

\item{slice_img1}{Slice image for filename1}

\item{slice_img2}{Slice image for filename2}
}
\value{
Total counts for all label numbers and slice numbers in both images,
overlap, only in slice_image1 and only in slice_image2.
}
\description{
Calculates total number of counts in both images and counts for
overlap, counts only in slice_image1, counts only in slice_image2.
}
\details{
Inputs one slice number and one label number. Function calculates
counts for total number of pixels in both images, pixels for overlap in the
given slice for the given label number, pixels only exist in slice_image1 and
pixels only exist in slice_image2. This function is primarily used by other
functions.
}
