% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractsliceno.R
\name{extractsliceno}
\alias{extractsliceno}
\title{Extract slice numbers}
\usage{
extractsliceno(filename, target_label, anatomical_plane = 3)
}
\arguments{
\item{filename}{One NIfTI-file}

\item{target_label}{One or more label numbers}

\item{anatomical_plane}{One for coronal plane, Two for sagittal plane and
Three for transverse plane. Default is Three.}
}
\value{
Slice numbers which include the target label number(s).
}
\description{
Extracts slice numbers for one or more supplied label
numbers.
}
\details{
Input one NIfTI-file and one or more target label numbers. Returns
slice number(s) which include the supplied target label number(s).
}
