% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_element.R
\name{files_element}
\alias{files_element}
\title{Total counts for filename}
\usage{
files_element(filename1, filename2)
}
\arguments{
\item{filename1}{A filename for one NIfTI-file}

\item{filename2}{A filename for one NIfTI-file}
}
\value{
Total number of counts in both images and counts for overlap, counts
only in image1, counts only in image2 for the given file pair.
}
\description{
Calculates total number of counts in both images and counts for
overlap, counts only in image1, counts only in image2 for given filename.
}
\details{
Function is used to calculate counts for calculating Dice
coefficients. This function is primarily used by other functions.
}
