% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelnaming2.R
\name{labelnaming2}
\alias{labelnaming2}
\title{Region volumes analysis data}
\usage{
labelnaming2(filename, labelnametibble)
}
\arguments{
\item{filename}{One or more NIfTI-files}

\item{labelnametibble}{A label name tibble}
}
\value{
Total volumes with label names, taken from the supplied label name
tibble, for all regions in the supplied NIfTI-file(s).
}
\description{
Calculates volumes for all regions and save the calculation
in a csv file.
}
\details{
Input one or more NIfTI-files and a label name tibble. Returns
total volumes, with label names, for all regions in the supplied NIfTI-file.
A label name tibble must have correct format and unique columns. First column
of the label name tibble must contain an integer label number and second
column must contain a label name string. Label numbers and label names must
be unique. Only regions which are in the label name tibble will be included
in the result.
}
