% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refmeanstd.R
\name{refmeanstd}
\alias{refmeanstd}
\title{Mean och std for z-score}
\usage{
refmeanstd(reflabels)
}
\arguments{
\item{reflabels}{At least two NIfTI-files}
}
\value{
Mean values and standard deviations for all regions in a reference
group.
}
\description{
Calculates mean and standard deviation for all regions in a
reference group.
}
\details{
Input NIfTI-files from a reference group to calculate mean and
standard deviation. The reference group must contain at least two NIfTI
files.
}
