% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regcount.R
\name{regcount}
\alias{regcount}
\title{Region counts}
\usage{
regcount(filename)
}
\arguments{
\item{filename}{One NIfTI-file}
}
\value{
Total counts and volumes for all region in a supplied nifty-file.
}
\description{
Calculates counts of all regions.
}
\details{
Inputs a NIfTI-file and returns total counts and volumes for all
regions.
}
