% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segcomparison.R
\name{segcomparison}
\alias{segcomparison}
\title{Segmentation comparison}
\usage{
segcomparison(
  filename_MR,
  filename_seg1,
  filename_seg2,
  Label_No,
  anatomical_plane = 3,
  rowno = 1,
  colno = 1
)
}
\arguments{
\item{filename_MR}{A NIfTI-file of an MR image.}

\item{filename_seg1}{First NIfTI-file of segmentation}

\item{filename_seg2}{Second NIfTI-file of segmentation}

\item{Label_No}{One label number for which region is to be displayed}

\item{anatomical_plane}{1 for coronal plane, 2 for sagittal plane and 3 for
transverse plane. Default is 3.}

\item{rowno}{Number of rows for showing images. Default is 1.}

\item{colno}{Number of columns for showing images. Default is 1.}
}
\value{
Label segmentation differences in supplied MR image for supplied
label number.
}
\description{
Displays label segmentation differences in an MR image.
}
\details{
Displays label segmentation differences in an MR image for a
supplied label number. Region only present in filename_seg 1 is displayed in
green color. Region only present in filename_seg2 is displayed in blue color.
Region present in both files is displayed in red color.
}
