% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segcomparison3D.R
\name{segcomparison3D}
\alias{segcomparison3D}
\title{Segmentation comparison in a 3D view}
\usage{
segcomparison3D(filename_MR, filename_seg1, filename_seg2, targetlabel)
}
\arguments{
\item{filename_MR}{A NIfTI-file of an MR image}

\item{filename_seg1}{First NIfTI-file of segmentation}

\item{filename_seg2}{Second NIfTI-file of segmentation}

\item{targetlabel}{One label number for which region is to be displayed}
}
\value{
3D view with segmentation differences.
}
\description{
Displays label segmentation differences in a 3D view.
}
\details{
Displays label segmentation differences in an MR image for supplied
label number. Region only present in filename_seg1 is displayed in green color.
Region only present in filename_seg2 is displayed in blue color. Region present in
both files is displayed in red color. 3D view can be rotated by dragging the
mouse.
}
