% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segcontour.R
\name{segcontour}
\alias{segcontour}
\title{Displays label contours in an MR image}
\usage{
segcontour(
  filename_MR,
  filename_seg,
  labelnametibble,
  slice_no,
  target_label,
  anatomical_plane = 3,
  rowno = 1,
  colno = 1
)
}
\arguments{
\item{filename_MR}{A NIfTI-file of an MR image}

\item{filename_seg}{A NIfTI-file of segmentation}

\item{labelnametibble}{A label name tibble with correct format and unique
columns. First column of the label name tibble must contain an integer label
number and second column must contain a label name string. Label numbers
and label names must be unique.}

\item{slice_no}{One or more slice numbers for which label contours are to be
displayed.}

\item{target_label}{One or more label numbers for which label contours are
to be displayed.}

\item{anatomical_plane}{1 for coronal plane, 2 for sagittal plane and 3 for
transverse plane. Default is 3.}

\item{rowno}{Number of rows for showing images. Default is 1.}

\item{colno}{Number of columns for showing images. Default is 1.}
}
\value{
One or more MR images with label contours for slice(s). A legend for
label contours with label names. Only contours which are in the label name
tibble and input label numbers will be displayed in the result.
}
\description{
Displays one or more label contours for one or more slices in
an MR image.
}
\details{
Displays one or more label contours for one or more slices in an MR
image for given label numbers. Contours lie on the edge of regions.
}
