% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segsurface3D.R
\name{segsurface3D}
\alias{segsurface3D}
\title{Displays label surfaces in a 3D view}
\usage{
segsurface3D(filename_MR, filename_seg, labelnametibble, targetlabel)
}
\arguments{
\item{filename_MR}{A NIfTI-file of an MR image}

\item{filename_seg}{A NIfTI-file of segmentation}

\item{labelnametibble}{A label name tibble with correct format and unique
columns. First column of the label name tibble should contain an integer
label number and second column should contain a label name string. Label
numbers and label names must be unique. Three columns, R, G, B, should be
included in the labelnametibble.}

\item{targetlabel}{One or more label numbers for which label surfaces are to
be shown.}
}
\value{
3D view with label surface(s). The different segments will have the
color from the R, G, B columns in the labelnametibble. A legend for label
surfaces with label names. Only surfaces which are in the label name tibble
and input label numbers will be displayed in the result.
}
\description{
Displays one or more label surfaces in an MR image in a 3D view.
}
\details{
Displays one or more label surfaces in an MR image for given label
numbers. 3D view can be rotated by dragging the mouse.
}
