% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_label.R
\name{slice_label}
\alias{slice_label}
\title{Extracts label numbers}
\usage{
slice_label(slice_no, slice_img1, slice_img2)
}
\arguments{
\item{slice_no}{One slice number with format integer}

\item{slice_img1}{Slice image for filename1}

\item{slice_img2}{Slice image for filename2}
}
\value{
Total counts for all label numbers and slice numbers in both images,
overlap, only in image1 and only in image2.
}
\description{
Extracts label numbers using slice.
}
\details{
Inputs one slice number. Returns a tibble. This function is
primarily used by other functions.
}
